refactor-优化测试用例,提高覆盖率

+++ src/main/java/com/example/demo/player/service/IPlayerServerBridging.java
@@ -30,6 +30,6 @@ public interface IPlayerServerBridging {
 	 * @param addValue 添加的经验
 	 * @return 添加结果回调
 	 */
-	CompletableFuture<Boolean> asyncAddPlayerExperience(int id, int addValue);
+	CompletableFuture<Boolean> asyncAddPlayerExperience(int id, long addValue);
 
 }
